IMPLEMENT FILTER COMPONENT
